UPDATE : unused module
import 되었지만 사용하지 않은 모듈 에 대한 코드 수정

+++ app/backend/src/users/users.service.ts
@@ -64,8 +64,8 @@ export class UsersService {
   }
 
   async deleteUsers(user_id: string){
-    // if (await this.usersRepo.count({user_id: user_id}) === 0)  // 존재하지 않은 유저이면
-      // return false;
+    if (await this.usersRepo.count({user_id: user_id}) === 0)  // 존재하지 않은 유저이면
+      return false;
     await this.achivementsService.deleteAllAchievements(user_id);
     await this.adminService.deleteAdmin(user_id);
     await this.banService.deleteBan(user_id);
